+++ src/scripts_paper/run_pretrainer_single_all.py
@@ -21,8 +21,8 @@ from src.modelzoo_nn import (
 )
 
 from utils import (
-    get_basin_id,
-    job_is_finished
+    check_finished_basins,
+    delete_unfinished_jobs,
 )
 
 # basin_file_all = '../../examples/4_basin_file.txt'
@@ -30,11 +30,16 @@ basin_file_all = '../../examples/569_basin_file.txt'
 
 nnmodel_type = 'lstm'   # 'lstm' or 'mlp'
 
-config_file = Path(f'config_run_nn_{nnmodel_type}_single.yml')
+# config_file = Path(f'config_run_nn_{nnmodel_type}_single.yml')
+# run_folder = f'runs_pretrainer_single_{nnmodel_type}'
+
+config_file = Path(f'config_run_nn_{nnmodel_type}_single270.yml')
 run_folder = f'runs_pretrainer_single_{nnmodel_type}270'
 
 MAX_WORKERS = 1
-CHECK_IF_FINISHED = False
+
+CHECK_IF_FINISHED = 1
+DELETE_IF_UNFINISHED = 0
 
 def train_model_for_basin(basin, config_file, project_path):
     '''
@@ -73,29 +78,6 @@ def train_model_for_basin(basin, config_file, project_path):
     # Train the model
     pretrainer.train()
 
-def check_finished_basins(runs_folder):
-
-    basin_finished = []
-    basin_unfinished = []
-    for folder in os.listdir(runs_folder):
-
-        basin = get_basin_id(folder)
-        if job_is_finished(script_path / runs_folder / folder):
-            basin_finished.append(str(int(basin)))
-        else:
-            basin_unfinished.append(basin)
-
-    return sorted(basin_finished), sorted(basin_unfinished)
-
-def delete_unfinished_jobs(runs_folder, basins):
-
-    # Find folders that contain basin in their name
-    folders = [f for f in os.listdir(runs_folder) if any(basin in f for basin in basins)]
-    
-    # Delete the folders
-    for folder in folders:
-        os.system(f'rm -rf {runs_folder / folder}')
-
 ####################################################################################################
 def main():
     
@@ -103,19 +85,27 @@ def main():
         with open(basin_file_all, 'r') as f:
             basins = f.readlines()
             
-        # max_workers = os.cpu_count()  # Adjust this based on your system and GPU availability
-        max_workers = MAX_WORKERS
+        # # max_workers = os.cpu_count()  # Adjust this based on your system and GPU availability
+        # max_workers = MAX_WORKERS
 
-        print(f"Training models for {len(basins)} basins using {max_workers} workers.")
+        # print(f"Training models for {len(basins)} basins using {max_workers} workers.")
 
         if CHECK_IF_FINISHED and os.path.exists(script_path / run_folder):
 
             # Find basins that are already finished and delete the unfinished jobs
-            basin_finished, basin_unfinished = check_finished_basins(run_folder)
-            delete_unfinished_jobs(script_path / run_folder, basin_unfinished)
+            basin_finished, basin_unfinished = check_finished_basins(script_path / run_folder)
+
+            print(f"Total basins: {len(basins)}")
+            print(f"Finished basins: {len(basin_finished)}")
+            print(f"Unfinished basins: {len(basin_unfinished)}", basin_unfinished)
+
+            if DELETE_IF_UNFINISHED:
+                delete_unfinished_jobs(script_path / run_folder, basin_unfinished)
 
             # Remove the finished basins from the list
             basins_to_continue = [basin for basin in basins if basin.strip() not in basin_finished]
+
+            print(f"Basins to continue: {len(basins_to_continue)}")
 
             # Update the list of basins
             basins = basins_to_continue
@@ -129,8 +119,8 @@ def main():
         # #         except Exception as e:
         # #             print(f"Training failed for a basin: {e}")
 
-        for basin in basins[280:]:
-            train_model_for_basin(basin, config_file, project_path)
+        # for basin in basins[:]:
+        #     train_model_for_basin(basin, config_file, project_path)
 
     else:
         print(f"File {basin_file_all} not found.")
